# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Analytics API for tracking and managing app data.",
        "title": "Minalytics API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/analytics/browsers": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves browser stats",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Analytics"
                ],
                "summary": "Get Browsers",
                "parameters": [
                    {
                        "type": "string",
                        "description": "app tracking ID",
                        "name": "trackingID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "start date",
                        "name": "startDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "end date",
                        "name": "endDate",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "stats fetched successfully",
                        "schema": {
                            "$ref": "#/definitions/main.BrowserResponse"
                        }
                    },
                    "400": {
                        "description": "invalid request paramaters",
                        "schema": {
                            "$ref": "#/definitions/main.APIStatus"
                        }
                    },
                    "500": {
                        "description": "failed to fetch browsers",
                        "schema": {
                            "$ref": "#/definitions/main.APIStatus"
                        }
                    }
                }
            }
        },
        "/analytics/countries": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves country stats",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Analytics"
                ],
                "summary": "Get Countries",
                "parameters": [
                    {
                        "type": "string",
                        "description": "app tracking ID",
                        "name": "trackingID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "start date",
                        "name": "startDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "end date",
                        "name": "endDate",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "stats fetched successfully",
                        "schema": {
                            "$ref": "#/definitions/main.CountryResponse"
                        }
                    },
                    "400": {
                        "description": "invalid request paramaters",
                        "schema": {
                            "$ref": "#/definitions/main.APIStatus"
                        }
                    },
                    "500": {
                        "description": "failed to fetch countries",
                        "schema": {
                            "$ref": "#/definitions/main.APIStatus"
                        }
                    }
                }
            }
        },
        "/analytics/devices": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves device stats",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Analytics"
                ],
                "summary": "Get Devices",
                "parameters": [
                    {
                        "type": "string",
                        "description": "app tracking ID",
                        "name": "trackingID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "start date",
                        "name": "startDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "end date",
                        "name": "endDate",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "stats fetched successfully",
                        "schema": {
                            "$ref": "#/definitions/main.DeviceResponse"
                        }
                    },
                    "400": {
                        "description": "invalid request paramaters",
                        "schema": {
                            "$ref": "#/definitions/main.APIStatus"
                        }
                    },
                    "500": {
                        "description": "failed to fetch devices",
                        "schema": {
                            "$ref": "#/definitions/main.APIStatus"
                        }
                    }
                }
            }
        },
        "/analytics/os": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves operating system stats",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Analytics"
                ],
                "summary": "Get OS",
                "parameters": [
                    {
                        "type": "string",
                        "description": "app tracking ID",
                        "name": "trackingID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "start date",
                        "name": "startDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "end date",
                        "name": "endDate",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "stats fetched successfully",
                        "schema": {
                            "$ref": "#/definitions/main.OSResponse"
                        }
                    },
                    "400": {
                        "description": "invalid request paramaters",
                        "schema": {
                            "$ref": "#/definitions/main.APIStatus"
                        }
                    },
                    "500": {
                        "description": "failed to fetch operating systems",
                        "schema": {
                            "$ref": "#/definitions/main.APIStatus"
                        }
                    }
                }
            }
        },
        "/analytics/pages": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves page stats",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Analytics"
                ],
                "summary": "Get Pages",
                "parameters": [
                    {
                        "type": "string",
                        "description": "app tracking ID",
                        "name": "trackingID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "start date",
                        "name": "startDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "end date",
                        "name": "endDate",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "stats fetched successfully",
                        "schema": {
                            "$ref": "#/definitions/main.PageResponse"
                        }
                    },
                    "400": {
                        "description": "invalid request paramaters",
                        "schema": {
                            "$ref": "#/definitions/main.APIStatus"
                        }
                    },
                    "500": {
                        "description": "failed to fetch pages",
                        "schema": {
                            "$ref": "#/definitions/main.APIStatus"
                        }
                    }
                }
            }
        },
...
        "/analytics/track": {
            "get": {
                "description": "Tracks an event based on encoded data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Analytics"
                ],
                "summary": "Track an event",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Base64 encoded event data",
                        "name": "data",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Event tracked successfully",
                        "schema": {
                            "$ref": "#/definitions/main.APIStatus"
                        }
                    },
                    "400": {
                        "description": "Invalid base64 or JSON data",
                        "schema": {
                            "$ref": "#/definitions/main.APIStatus"
                        }
                    },
                    "500": {
                        "description": "Failed to resolve geolocation or track event",
                        "schema": {
                            "$ref": "#/definitions/main.APIStatus"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "main.APIStatus": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "main.BrowserResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/main.BrowserStats"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "main.BrowserStats": {
            "type": "object",
            "properties": {
                "browser": {
                    "type": "string"
                },
                "percentage": {
                    "type": "integer"
                }
            }
        },
        "main.CountryResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/main.CountryStats"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "main.DeviceResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/main.DeviceStats"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "main.OSResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/main.OSStats"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "main.PageResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/main.PageStats"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}